+++ .gitignore
@@ -9,4 +9,5 @@
 *.out
 *out
 *.idea
+*target
 9
