+++ src/main/kotlin/main/func/TestingKotlinMain.kt
@@ -1,0 +1,10 @@
+package main.kotlin.main.func
+
+import org.springframework.boot.runApplication
+
+@org.springframework.boot.autoconfigure.SpringBootApplication
+class TestingKotlinMain {
+    fun main(args: Array<String>) {
+        runApplication<TestingKotlinMain>(*args)
+    }
+}
