+++ src/main/kotlin/main/func/conf/Check.kt
@@ -1,0 +1,9 @@
+package main.func.conf
+
+fun main(args: Array<String>) {
+    area(10, 20, 30)
+}
+
+fun area(length: Int, width: Int, height: Int) {
+    println("length is $length, width is $width and height is $height")
+}
