+++ src/main/kotlin/main/func/loops/For.kt
@@ -1,0 +1,4 @@
+package main.func.loops
+
+class For {
+}
